Worked on new build system, fixed bug
Fully converted the x86 build code to using the
new makefile format. Everything builds and runs fine.

In the course of doing this, I found that the "-kernel"
switch to qemu was not properly simulating GRUB; qemu
simply booted a kernel with NO MULTIBOOT HEADER. In addition,
a real grub bootloader uses the VIRTUAL address of the start
function; since the linker script used was linking the start
function to somewhere over 0xc0000000, GRUB complained that
the start address was not within a segment.